% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_calculo.R
\name{agua_server}
\alias{agua_server}
\title{Agua Server}
\usage{
agua_server(id, app_state, parent)
}
\arguments{
\item{app_state}{estado da aplicaçào}

\item{parent}{shiny parent input}

\item{input}{shiny input}
}
\value{
shiny server
}
\description{
Agua Server
}
