% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_agua.R
\name{agua_ui}
\alias{agua_ui}
\title{Gera interface para componente de água}
\usage{
agua_ui(id, app_state)
}
\arguments{
\item{ns}{é o namespace do módulo de interface gráfica}
}
\value{
o html da interface gráfica
}
\description{
Gera interface para componente de água
}
