% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_config.R
\name{anos_disponiveis_base_calculo}
\alias{anos_disponiveis_base_calculo}
\title{Retorna os anos disponíveis arquivos da base de cáculo no formate \code{componente_source_YYYY.csv}}
\usage{
anos_disponiveis_base_calculo(componente, source)
}
\arguments{
\item{componente}{Nome do componente (ex: "agua", "esgoto")}

\item{source}{Nome da fonte de dados (ex: "snis", "sinisa")}
}
\value{
Vetor de anos disponíveis encontrados nos arquivos CSV
}
\description{
Retorna os anos disponíveis arquivos da base de cáculo no formate \code{componente_source_YYYY.csv}
}
