% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_config.R
\name{atendimento_anos_disponiveis}
\alias{atendimento_anos_disponiveis}
\title{Retorna os anos disponíveis para atendimento de acordo com a fonte}
\usage{
atendimento_anos_disponiveis(componente, fonte_nome)
}
\arguments{
\item{fonte_nome}{Nome da fonte de dados (ex: "censo", "sinisa", "pnadc")}
}
\value{
Vetor de anos disponíveis para a fonte informada
}
\description{
Retorna os anos disponíveis para atendimento de acordo com a fonte
}
