% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_config.R
\name{config_ui}
\alias{config_ui}
\title{Gera interface para a configuração das fontes de dados}
\usage{
config_ui(id, app_state)
}
\arguments{
\item{ns}{é o namespace do módulo de interface gráfica}
}
\value{
o html da interface gráfica
}
\description{
Gera interface para a configuração das fontes de dados
}
