% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_dashboard.R
\name{dash_residuos}
\alias{dash_residuos}
\title{Title}
\usage{
dash_residuos(ns)
}
\arguments{
\item{ns}{é o namespace do módulo de interface gráfica}
}
\value{
o html da interface gráfica
}
\description{
Title
}
