% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_esgoto.R
\name{esgoto_ui}
\alias{esgoto_ui}
\title{Title}
\usage{
esgoto_ui(id, app_state)
}
\arguments{
\item{id}{é o namespace do módulo de interface gráfica}
}
\value{
o html da interface gráfica
}
\description{
Title
}
