% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_config.R
\name{estrutura_anos_disponiveis}
\alias{estrutura_anos_disponiveis}
\title{Retorna os anos disponíveis para um componente e fonte de dados estruturais}
\usage{
estrutura_anos_disponiveis(componente, source)
}
\arguments{
\item{componente}{Nome do componente (ex: "agua", "esgoto")}

\item{source}{Nome da fonte de dados (ex: "snis", "sinisa")}
}
\value{
Vetor de anos disponíveis encontrados nos arquivos CSV
}
\description{
Retorna os anos disponíveis para um componente e fonte de dados estruturais
}
