% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_projecao.R
\name{get_fonte1_list}
\alias{get_fonte1_list}
\title{Retorna lista de opções de dados ibge censo}
\usage{
get_fonte1_list()
}
\value{
um vetor com os nomes
}
\description{
Retorna lista de opções de dados ibge censo
}
