% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_projecao.R
\name{get_fonte2_list}
\alias{get_fonte2_list}
\title{Retorna lista de opções de dados ibge censo e estimativa}
\usage{
get_fonte2_list()
}
\value{
um vetor com os nomes
}
\description{
Retorna lista de opções de dados ibge censo e estimativa
}
