% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_config.R
\name{get_sinapi_list}
\alias{get_sinapi_list}
\title{Retorna lista de opções de dados do SINAPI}
\usage{
get_sinapi_list()
}
\value{
um vetor com os nomes
}
\description{
Retorna lista de opções de dados do SINAPI
}
