% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_calculo.R
\name{modulo_calculo}
\alias{modulo_calculo}
\title{Modulo calculo}
\usage{
modulo_calculo(id, app_state, parent)
}
\arguments{
\item{app_state}{estado da aplicaçào}

\item{parent}{shiny parent input}

\item{input}{shiny input}
}
\value{
shiny server
}
\description{
Modulo calculo
}
