% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_parametros.R
\name{parameters_server}
\alias{parameters_server}
\title{Modulo calculo}
\usage{
parameters_server(id, parent_session)
}
\arguments{
\item{input}{shiny input}

\item{app_state}{estado da aplicaçào}
}
\value{
shiny server
}
\description{
Modulo calculo
}
