% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_parametros.R
\name{parametros_ui}
\alias{parametros_ui}
\title{Gera interface para os parâmetros de cálculo}
\usage{
parametros_ui(id, app_state)
}
\arguments{
\item{ns}{é o namespace do módulo de interface gráfica}
}
\value{
o html da interface gráfica
}
\description{
Gera interface para os parâmetros de cálculo
}
