% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_projecao.R
\name{projecao_populacional_ui}
\alias{projecao_populacional_ui}
\title{Title}
\usage{
projecao_populacional_ui(id, app_state)
}
\arguments{
\item{ns}{é o namespace do módulo de interface gráfica}
}
\value{
o html da interface gráfica
}
\description{
Title
}
