% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_residuos.R
\name{residuos_ui}
\alias{residuos_ui}
\title{Title}
\usage{
residuos_ui(id, app_state)
}
\arguments{
\item{ns}{é o namespace do módulo de interface gráfica}
}
\value{
o html da interface gráfica
}
\description{
Title
}
